add deploy --up

+++ bundles/deploy.py
@@ -24,27 +24,28 @@ def copy_project_into_volume(copy_target):
             entrypoint='sh'
             )
 
-def deploy(mode):
+def deploy(mode, onlyUp = False):
     ADD_DEPLOY_CONFIG = ' -f docker-compose.yml.deploy -f docker-compose.yml.%s '%mode
     def call_docker_compose_deploy(command):
         return os.system(env.docker_compose(ADD_DEPLOY_CONFIG + command))
     def call_docker_compose_deploy_new(command):
         return os.system('NEW=_new ' + env.docker_compose(ADD_DEPLOY_CONFIG + command))
     run.init_volumes()
-    if 0 != call_docker_compose_deploy_new(env.run(["cd %(output_path)s_new && %(prepare)s"\
-            %{'prepare': (run.COMMAND_DEPENDENCES + ' && ' + run.COMMAND_PREPARE), 'output_path': VOLUME_PATH}])):
-        return
-    call_docker_compose_deploy_new(env.down())
-    if 0 != call_docker_compose_deploy_new(env.run(["cd .. &&\
-            rm -rf %(output_path)s && mv %(output_path)s_new %(output_path)s"\
-            %{'output_path': VOLUME_PATH}])):
-        return
-    if 0 != call_docker_compose_deploy(env.run(["cd .. &&\
-            rm -rf %(output_path)s_new"\
-            %{'output_path': VOLUME_PATH}])):
-        return
-    if 0 != call_docker_compose_deploy(env.run([run.COMMAND_DB_MIGRATE])):
-        return
+    if not onlyUp:
+        if 0 != call_docker_compose_deploy_new(env.run(["cd %(output_path)s_new && %(prepare)s"\
+                %{'prepare': (run.COMMAND_DEPENDENCES + ' && ' + run.COMMAND_PREPARE), 'output_path': VOLUME_PATH}])):
+            return
+        call_docker_compose_deploy_new(env.down())
+        if 0 != call_docker_compose_deploy_new(env.run(["cd .. &&\
+                rm -rf %(output_path)s && mv %(output_path)s_new %(output_path)s"\
+                %{'output_path': VOLUME_PATH}])):
+            return
+        if 0 != call_docker_compose_deploy(env.run(["cd .. &&\
+                rm -rf %(output_path)s_new"\
+                %{'output_path': VOLUME_PATH}])):
+            return
+        if 0 != call_docker_compose_deploy(env.run([run.COMMAND_DB_MIGRATE])):
+            return
     call_docker_compose_deploy(env.up())
 
 def show_deploy_path(path):
@@ -54,13 +55,13 @@ def show_deploy_path(path):
 
 def deploy_local(args = []):
     copy_project_into_volume(os.path.realpath('./deploy'))
-    deploy(env.env)
+    deploy(env.env, len(args) > 0 and args[0] == '--up')
     show_deploy_path(os.path.realpath('.' + VOLUME_PATH))
 
 def deploy_nginx(args = []):
     copy_project_into_volume('nginx-deploy')
-    deploy('nginx')
-    show_deploy_path('nginx-deploy:' + VOLUME_PATH)
+    deploy('nginx', len(args) > 0 and args[0] == '--up')
+    show_deploy_path(VOLUME_PATH)
 
 _actions = {}
 if env.env == 'staging' or env.env == 'production':
@@ -69,15 +70,18 @@ if env.env == 'staging' or env.env == 'production':
             'desc': """Deploy into docker as Server.
                                   Before use it:
                                   * Put config file into docker/env/${ENV}/
-                                  * -e BRANCH=<YOUR DEPLOY GIT BRANCH>""",
+                                  * -e BRANCH=<YOUR DEPLOY GIT BRANCH>
+                                  * --up: only up server""",
             'action': deploy_local
         },
         'rails:deploy:nginx': {
             'desc': """Deploy into docker as Server, Serve by nginx in docker.
                                   Before use it:
                                   * Put config file into docker/env/${ENV}/
-                                  * -e BRANCH=<YOUR DEPLOY GIT BRANCH>""",
+                                  * -e BRANCH=<YOUR DEPLOY GIT BRANCH>
+                                  * --up: only up server""",
             'action': deploy_nginx
         }
     }
+
